fix some bugs

+++ src/camera.cpp
@@ -196,10 +196,10 @@ Ray Camera::generate_ray_for_thin_lens(double x, double y, double rndR, double r
 }
 
 void Camera::kernel_struct(kernel_camera_t *kernel_camera) {
-  kernel_camera->h_fov = radians(hFov);
-  kernel_camera->v_fov = radians(vFov);
-  kernel_camera->n_clip = nClip;
-  kernel_camera->f_clip = fClip;
+  kernel_camera->h_fov = (float) radians(hFov);
+  kernel_camera->v_fov = (float) radians(vFov);
+  kernel_camera->n_clip = (float) nClip;
+  kernel_camera->f_clip = (float) fClip;
   kernel_camera->pos = cglVectorToKernel(pos);
   kernel_camera->c2w = cglMatrixToKernel(c2w);
 }
